refactor: remove reducer
